Added SwipeModel, MatchModel
Added other users fields such as age, bio etc.
Added new routes including "/users" to get all users, "/update-user" to update the user record, "/add-swipe" to record the swipe and check for matches, "/matches" which gets all the matches for a user.
Also added a send_notification() function which sends emails to the users who are matched.
The routes are untested because I am unsure about the schema of the database and we need to discuss this.